+++ external/llvm-project/libc/src/ctype/isdigit.cpp
@@ -8,14 +8,12 @@
 
 #include "src/ctype/isdigit.h"
 #include "src/__support/common.h"
-#include "src/__support/ctype_utils.h"
+#include "src/ctype/ctype_utils.h"
 
 namespace __llvm_libc {
 
 // TODO: Currently restricted to default locale.
 // These should be extended using locale information.
-LLVM_LIBC_FUNCTION(int, isdigit, (int c)) {
-  return static_cast<int>(internal::isdigit(static_cast<unsigned>(c)));
-}
+LLVM_LIBC_FUNCTION(int, isdigit, (int c)) { return internal::isdigit(c); }
 
 } // namespace __llvm_libc
