+++ external/llvm-project/libc/src/ctype/isgraph.cpp
@@ -9,14 +9,12 @@
 #include "src/ctype/isgraph.h"
 
 #include "src/__support/common.h"
-#include "src/__support/ctype_utils.h"
+#include "src/ctype/ctype_utils.h"
 
 namespace __llvm_libc {
 
 // TODO: Currently restricted to default locale.
 // These should be extended using locale information.
-LLVM_LIBC_FUNCTION(int, isgraph, (int c)) {
-  return static_cast<int>(internal::isgraph(static_cast<unsigned>(c)));
-}
+LLVM_LIBC_FUNCTION(int, isgraph, (int c)) { return internal::isgraph(c); }
 
 } // namespace __llvm_libc
