+++ external/llvm-project/libc/src/ctype/islower.cpp
@@ -7,7 +7,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "src/ctype/islower.h"
-#include "src/__support/ctype_utils.h"
+#include "src/ctype/ctype_utils.h"
 
 #include "src/__support/common.h"
 
@@ -15,8 +15,6 @@ namespace __llvm_libc {
 
 // TODO: Currently restricted to default locale.
 // These should be extended using locale information.
-LLVM_LIBC_FUNCTION(int, islower, (int c)) {
-  return static_cast<int>(internal::islower(static_cast<unsigned>(c)));
-}
+LLVM_LIBC_FUNCTION(int, islower, (int c)) { return internal::islower(c); }
 
 } // namespace __llvm_libc
